% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_directory.R
\name{source_directory}
\alias{source_directory}
\title{Source R Files in Directory}
\usage{
source_directory(dir_name, verbose = FALSE, show_warnings = TRUE)
}
\arguments{
\item{dir_name}{String name for directory within current working directory.}

\item{verbose}{Produce verbose output? Defaults to FALSE.}

\item{show_warnings}{Show warning messages, if any? Defaults to TRUE.}
}
\description{
\code{source_directory} sources all R files within a directory.
}
\examples{
source_directory("src")
}
